Refactor compose structure
