+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/Router.kt
@@ -3,9 +3,9 @@ package de.hsfl.budgetBinder.compose
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.rememberCoroutineScope
-import de.hsfl.budgetBinder.compose.screens.dashboard.DashboardComponent
-import de.hsfl.budgetBinder.compose.screens.login.LoginComponent
-import de.hsfl.budgetBinder.compose.screens.register.RegisterComponent
+import de.hsfl.budgetBinder.screens.dashboard.DashboardComponent
+import de.hsfl.budgetBinder.screens.login.LoginComponent
+import de.hsfl.budgetBinder.screens.register.RegisterComponent
 import de.hsfl.budgetBinder.presentation.flow.RouterFlow
 import de.hsfl.budgetBinder.presentation.Screen
 import org.kodein.di.instance
