+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/textfield/EmailTextField.kt
@@ -1,0 +1,45 @@
+package de.hsfl.budgetBinder.compose.textfield
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.OutlinedTextField
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun EmailTextField(
+    modifier: Modifier = Modifier,
+    value: String,
+    onValueChange: (String) -> Unit,
+    label: @Composable (() -> Unit)? = null,
+    enabled: Boolean = true,
+    isError: Boolean = false
+) {
+    Column {
+        OutlinedTextField(
+            modifier = modifier,
+            value = value,
+            onValueChange = onValueChange,
+            enabled = enabled,
+            isError = isError,
+            label = label,
+            singleLine = true,
+            trailingIcon = {
+                if (isError) {
+                    Icon(Icons.Default.Info, contentDescription = null)
+                }
+            }
+        )
+    }
+    if (isError) {
+        Text(
+            text = "Please enter a valide email address",
+            color = MaterialTheme.colors.error,
+            style = MaterialTheme.typography.subtitle2
+        )
+    }
+}
